requirements.txt
streamlit
ultralytics
Pillow
easyocr